Controllers

+++ src/main/java/com/igor/contacts/service/PersonService.java
@@ -14,6 +14,9 @@ public class PersonService {
     @Autowired
     private PersonRepository personRepository;
 
+    @Autowired
+    private ContactService contactService;
+
 
     public Optional<Person> findById(Long id){
 
@@ -23,7 +26,7 @@ public class PersonService {
 
     public List<Person> searchPersons(Optional<String> query){
 
-        //Checa se foi passado o parametro de filtragem para busca, se sim busca pelo nome, e retorna todos os registros caso contrário
+        //Checa se foi passado o parametro de filtragem para busca, se sim busca pelo nome, e retorna todos os registros caso contrário - Igor
         return query.isPresent() ? personRepository.findAllByNameContaining(query.get()) : personRepository.findAll();
 
     }
@@ -36,14 +39,19 @@ public class PersonService {
 
             if(dbPerson.isPresent()){
 
-                dbPerson.get().setName(person.getName());
+                if(person.getName() != null && !person.getName().isEmpty()){
+                    dbPerson.get().setName(person.getName());
+                }
+
+                //Se algum dos contatos tiver uma informação para atualizar, executa a atualizaçãp
+                person.getContacts().forEach(c -> contactService.save(c,c.getPerson()));
 
                 return personRepository.save(dbPerson.get());
 
             }
 
             //Se por alguma razão o id que foi passado no objeto não existir no banco de dados, apenas remove o id inválido
-            //e salva o objeto como novo registro
+            //e salva o objeto como novo registro - Igor
             person.setId(null);
 
 
@@ -55,7 +63,19 @@ public class PersonService {
 
     public void delete(Long id){
 
-        personRepository.deleteById(id);
+        Optional<Person> person = findById(id);
+
+        if(person.isPresent()){
+
+            //Se a pessoa tiver contatos associados a ela, precisa deletar os contatos primeiro.
+            //Caso contrário o banco não deixará deletar por causa da chave estrangeira e da constraint - Igor
+            person.get().getContacts().forEach(c -> contactService.delete(c.getId()));
+
+            personRepository.deleteById(id);
+
+        }
+
+
 
     }
